Change ConfirmCompositionText() from uint32_t to size_t.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3783487
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1028922}

+++ ui/base/ime/fake_text_input_client.h
@@ -37,7 +37,7 @@ class FakeTextInputClient : public TextInputClient {
 
   // TextInputClient:
   void SetCompositionText(const CompositionText& composition) override;
-  uint32_t ConfirmCompositionText(bool keep_selection) override;
+  size_t ConfirmCompositionText(bool keep_selection) override;
   void ClearCompositionText() override;
   void InsertText(
       const std::u16string& text,
